# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Skew AI — Voice-of-Customer Intelligence Platform
  description: 'Skew AI (skewai): domain-agnostic voice-of-customer intelligence.
    Plug in a Domain Pack and get an AI voice agent, a swarm of triage/RCA agents,
    and the Qubot v2 auditor.


    ## Auth

    When `FRONTLINE_API_KEY` is set, send `X-API-Key` or `Authorization: Bearer`.


    ## Versioning

    Current API version **1** (header `API-Version`). Paths under `/api/...` are v1
    without a URL prefix. Breaking changes will introduce `/api/v2/...`.


    ## Errors

    Errors use RFC 7807-shaped JSON (`type`, `title`, `status`, `detail`) while keeping
    FastAPI''s `detail` string for existing clients.


    ## Pagination

    Collection endpoints accept `limit`, `offset`, and optional `cursor`, and return
    a `pagination` object (`has_more`, `next_offset`, `next_cursor`).'
  version: 2.0.0
paths:
...
  /api/frontline/enterprise/interactions/recent:
    get:
      tags:
      - enterprise
      summary: Recent Interactions
      description: Newest interactions for Enterprise Ops pickers (additive catalog).
      operationId: recent_interactions_api_frontline_enterprise_interactions_recent_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Recent Interactions Api Frontline Enterprise Interactions
                  Recent Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/timeline/{interaction_id}:
    get:
      tags:
      - enterprise
      summary: Get Timeline
      operationId: get_timeline_api_frontline_enterprise_timeline__interaction_id__get
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Timeline Api Frontline Enterprise Timeline  Interaction
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/root-cause/{interaction_id}:
    get:
      tags:
      - enterprise
      summary: Get Root Cause
      operationId: get_root_cause_api_frontline_enterprise_root_cause__interaction_id__get
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Root Cause Api Frontline Enterprise Root Cause  Interaction
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/root-cause:
    get:
      tags:
      - enterprise
      summary: List Root Causes
      operationId: list_root_causes_api_frontline_enterprise_root_cause_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Root Causes Api Frontline Enterprise Root Cause
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/copilot:
    post:
      tags:
      - enterprise
      summary: Copilot Query
      operationId: copilot_query_api_frontline_enterprise_copilot_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Copilot Query Api Frontline Enterprise Copilot Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/risk/active:
    get:
      tags:
      - enterprise
      summary: Risk Active
      operationId: risk_active_api_frontline_enterprise_risk_active_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: persist
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Persist
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Risk Active Api Frontline Enterprise Risk Active Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/risk/{interaction_id}/history:
    get:
      tags:
      - enterprise
      summary: Risk History
      operationId: risk_history_api_frontline_enterprise_risk__interaction_id__history_get
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Risk History Api Frontline Enterprise Risk  Interaction
                  Id  History Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/risk/{interaction_id}:
    get:
      tags:
      - enterprise
      summary: Risk One
      operationId: risk_one_api_frontline_enterprise_risk__interaction_id__get
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: persist
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Persist
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Risk One Api Frontline Enterprise Risk  Interaction
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/graph:
    get:
      tags:
      - enterprise
      summary: Ops Graph
      operationId: ops_graph_api_frontline_enterprise_graph_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: limit_cases
        in: query
        required: false
        schema:
          type: integer
          default: 40
          title: Limit Cases
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Ops Graph Api Frontline Enterprise Graph Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/memory:
    get:
      tags:
      - enterprise
      summary: Memory List
      operationId: memory_list_api_frontline_enterprise_memory_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Memory List Api Frontline Enterprise Memory Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/memory/lookup:
    get:
      tags:
      - enterprise
      summary: Memory Lookup
      operationId: memory_lookup_api_frontline_enterprise_memory_lookup_get
      parameters:
      - name: pack_id
        in: query
        required: true
        schema:
          type: string
          title: Pack Id
      - name: entity_1
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity 1
      - name: entity_2
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity 2
      - name: entity_3
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity 3
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Memory Lookup Api Frontline Enterprise Memory Lookup
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/memory/upsert/{interaction_id}:
    post:
      tags:
      - enterprise
      summary: Memory Upsert
      operationId: memory_upsert_api_frontline_enterprise_memory_upsert__interaction_id__post
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Memory Upsert Api Frontline Enterprise Memory Upsert  Interaction
                  Id  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/scenarios:
    get:
      tags:
      - enterprise
      summary: Scenarios List
      operationId: scenarios_list_api_frontline_enterprise_scenarios_get
      parameters:
      - name: pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pack Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Scenarios List Api Frontline Enterprise Scenarios
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - enterprise
      summary: Scenarios Create
      operationId: scenarios_create_api_frontline_enterprise_scenarios_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Scenarios Create Api Frontline Enterprise Scenarios
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/frontline/enterprise/scenarios/validate:
    post:
      tags:
      - enterprise
      summary: Scenarios Validate
      description: Validate playbook steps without saving (pure; no orchestrator).
      operationId: scenarios_validate_api_frontline_enterprise_scenarios_validate_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Scenarios Validate Api Frontline Enterprise Scenarios
                  Validate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/frontline/enterprise/scenarios/{scenario_id}/run:
    post:
      tags:
      - enterprise
      summary: Scenarios Run
      description: Run a scenario off the event loop (same spirit as /simulate + to_thread).
      operationId: scenarios_run_api_frontline_enterprise_scenarios__scenario_id__run_post
      parameters:
      - name: scenario_id
        in: path
        required: true
        schema:
          type: string
          title: Scenario Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Scenarios Run Api Frontline Enterprise Scenarios  Scenario
                  Id  Run Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...